@@ -1,8 +1,12 @@
-/* eslint @typescript-eslint/no-var-requires: "off" */
-const { getGitVersion, getCommitHash } = require('./scripts/gitInfo');
-const { dependencies } = require('./package.json');
-const path = require('node:path');
-const printer = require('./scripts/printer').printer;
+import { getGitVersion, getCommitHash } from './scripts/gitInfo.mjs';
+import fs from 'node:fs';
+import path from 'node:path';
+import { printer } from './scripts/printer.mjs';
+import debugLocal from 'next-debug-local';
+
+const dependencies = JSON.parse(fs.readFileSync('./package.json', 'utf8'))[
+  'dependencies'
+];
 
 const enableDebugLocal = path.isAbsolute(process.env.LOCAL_BLOCK_SUITE ?? '');
 const EDITOR_VERSION = enableDebugLocal
@@ -56,7 +60,6 @@ const nextConfig = {
   ],
   webpack: config => {
     config.experiments = { ...config.experiments, topLevelAwait: true };
-    config.resolve.alias['yjs'] = require.resolve('yjs');
     config.module.rules.push({
       test: /\.md$/i,
       loader: 'raw-loader',
@@ -81,7 +84,7 @@ const nextConfig = {
 };
 
 const baseDir = process.env.LOCAL_BLOCK_SUITE ?? '/';
-const withDebugLocal = require('next-debug-local')(
+const withDebugLocal = debugLocal(
   {
     '@blocksuite/editor': path.resolve(baseDir, 'packages', 'editor'),
     '@blocksuite/blocks/models': path.resolve(
@@ -115,4 +118,4 @@ const detectFirebaseConfig = () => {
 };
 detectFirebaseConfig();
 
-module.exports = withDebugLocal(nextConfig);
+export default withDebugLocal(nextConfig);
