@@ -1,8 +1,5 @@
 // @ts-check
-
-// import { execSync } from 'child_process'
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { execSync } = require('child_process');
+import { execSync } from 'node:child_process';
 
 const hasGit = () => {
   try {
@@ -41,7 +38,4 @@ const getGitVersion = () => {
 const getCommitHash = (rev = 'HEAD') =>
   execSync(`git rev-parse --short ${rev}`).toString();
 
-module.exports = {
-  getGitVersion,
-  getCommitHash,
-};
+export { getCommitHash, getGitVersion };
