@@ -1,6 +1,5 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const chalk = require('chalk');
-const printer = {
+import chalk from 'chalk';
+export const printer = {
   debug: msg => {
     const result = chalk.green`debug` + chalk.white('  - ' + msg);
     console.log(result);
@@ -17,5 +16,3 @@ const printer = {
     return result;
   },
 };
-
-module.exports = { printer };
