@@ -15,17 +15,16 @@ interface LoginModalProps {
   workspaceId: string;
   onInviteSuccess: () => void;
 }
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
+
 export const debounce = <T extends (...args: any) => any>(
   fn: T,
   time?: number,
   immediate?: boolean
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ): ((...args: any) => any) => {
   let timeoutId: null | number;
   let defaultImmediate = immediate || false;
   const delay = time || 300;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+
   return (...args: any) => {
     if (defaultImmediate) {
       fn.apply(this, args);
@@ -35,7 +34,7 @@ export const debounce = <T extends (...args: any) => any>(
     if (timeoutId) {
       clearTimeout(timeoutId);
     }
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+
     // @ts-ignore
     timeoutId = setTimeout(() => {
       fn.apply(this, args);
